Maintain per-wiki IDs for the logged-in user
A user may edit a page for a wiki other than the app's current language
state.  We should log the appropriate ID for the wiki.  Currently we
use the id for the app language state.

This updates user ID handling to maintain a map of the languages with
which the user interacts, so that the appropriate one may be logged when
the user makes an edit.

Change-Id: Ia86803be5c861f7e2dbded43f691dbc13b0c7fbb