Prevent user attempting to create a duplicate-titled reading list
Checks the title in the reading list creation dialog against the list
of existing list titles, disabling the positive button and displaying
error text in the case of a match.

My personal preference and sense of what a user intends favor this
solution over something like creating a new list and  appending a " (1)"
to the title.

Bug: T156017
Change-Id: I8a46db35724104434053a802408ffc58c4d800b7