Remove 'edit history' action from revert notification.
Per Design recommendation:
User testing has revealed some confusion about the 'view edit history'
button that we show in the Revert notification. The mobile web
presentation of the edit history is not yet friendly enough, and can make
some users believe that they broke the internet.

Let's remove this button for now, and just keep the functionality of the
notification taking the user back to the last-edited article, which will
then show our "We know you tried" dialog, which still contains a link to
the edit history, should they choose to click it.

Change-Id: I24a66432102aa08d085a95199b3e2207ade57a61