Hygiene: Replace usages of HashSet with ArraySet
Google bills ArraySet, which is included in the Android Support Library as
of 25.1.0, as a generic set data structure that is designed to be more
memory efficient than a traditional HashSet.

Note: the find-and-replace one-liner used to make this patch had the side-
effect of adding a newline to the end of every Java file in the repo that
didn't have one.

Change-Id: I839a8d253823df4b0ceea4396c64fb0fd5f538d1