Fix reading list sorting
Moves the list page sorting logic into ReadingListDetailView where it can
interact with the displayedPages list that controls what's actually
displayed to the user since the introduction of search filtering.

This is a better place for the sorting logic to live anyway since it's an
entirely presentational concern.  Some needless sorting in the ReadingList
model is eliminated.

Bug: T151662
Change-Id: I05f0f49a9021dfecc587ca51091593ddba9b9c18