Improve implicit re-login behavior when editing descriptions.
This adds some logic to the base class MwPostResponse that lets us know
when the user needs to be re-logged-in automatically, before we retry
fetching the edit token and attempting the edit again. This is in contrast
with our current behavior of explicitly re-logging-in the user upon every
single edit.

This also updates the retry logic when saving a description, by resetting
the retry counter to 0 when the Save button is clicked. (Previously the
retry counter was never reset.)

Note: 2FA users will still need to go to the Login activity to log in
correctly, prior to making additional edits.

Change-Id: I8e2a3f8997ae529798d19124a81728206836693e