Correctly show success snackbar upon logging in.
It looks like we have the logic to show a Snackbar when the user logs in
successfully, but it was never being called. This is because the
onActivityResult() method gets called at the same nesting level from which
startActivityForResult() was called. (And we were invoking LoginActivity
from one level deeper than the level where we were catching
onActivityResult).

This moves the call to invoke the Login activity up a level, making the
invocation symmetric with the expectation of receiving onActivityResult().

Change-Id: I54c3730ae4af0cbfb3f42c36dc7acff96b969f34