Drop saved pages database table
This has been superseded by the reading list page table(s) and can be
dropped.

Some code is retained to preserve sequential migrations.

Additional code relating to the old saved pages implementation will be
removed in a separate, provisional patch, which shouldn't be merged until
the updated implementation is in place.

Bug: T135401
Change-Id: I869a88d0833975ae908a1a7470d1715df27bd0c6