Prevent user from creating a list with an empty title
It's currently possible for a user to create a reading list with an empty
string as the title.  When this happens, "Untitled" will be displayed as
the list title, but the true title will be an empty string, and pages will
disappear from the reading list almost immediately.

Let's prevent users from creating reading lists with empty titles.

Change-Id: If330328a60b60db96f5ccaa7af2237bcbc65a37d