[Facepalm] Load CSS as Unicode, not ASCII.
Our CSS contains several "#content" directives that inject additional
content into the HTML, some of which contain Unicode characters. We've
been noticing that these characters don't seem to be rendering correctly
in our WebView.

It turns out that the WebView has been loading the CSS content as ASCII,
and we need to explicitly declare our content to be UTF-8 in a "meta" tag
in the HTML. It's surprising that UTF-8 is not the default, but there
you go.

Bug: T158067
Change-Id: Ia0cd07d99ba1fb09112a883901c1d5d40a29eba7