Improve symmetry of setting/nulling callbacks on feed views.
So, here's the thing:
When feed views get destroyed, we need to explicitly null out the
"callback" field that they contain. Otherwise, it's known to cause memory
leak(s) on certain APIs.

So far, we've been using the onViewAttachedToWindow() method to give the
callback to the view, and similarly using onViewDetachedFromWindow() to
set the callback to null, which seems to work well in preventing memory
leaks.

I've noticed, however, that in certain feed views, we have *nested*
RecyclerViews, and in the adapters of those nested RecyclerViews, we don't
actually use the attached/detached-from-window logic for setting and
nulling the callback (instead we just set the callback in onBindViewHolder).

My guess is that the latest version of appcompat has exposed this
inconsistency, in the form of race condition between the parent and nested
RecyclerAdapters in setting the callback.

This patch updates our "child" RecyclerAdapters to follow the same pattern
of setting and nulling the callback based on the attached/detached calls.

Bug: T158347
Change-Id: Ieb5461dbee0627df5252c51c54817627eedcf4a5