[Facepalm?] Correctly pass CentralAuth cookies to Wikidata.
When the user logs in through the app, we receive a CentralAuth cookie
which has a domainSpec of ".wikipedia.org".

Then, when making requests to edit Wikidata descriptions, we pass that
same cookie to the wikidata.org domain. However, our code that determines
whether to pass the cookie was checking if the domainSpec *equals*
wikipedia.org, rather than *ends with* wikipedia.org, so the cookie wasn't
getting passed, and the edit not going through successfully.

I'm... not sure how/why this worked before. Did something change on the
server side?

Change-Id: I91c73238467d3d54bbf06373fb0534baace9c438