Improve/simplify bottom tab bar behavior.
This moves away from using a fake AppBarLayout to control the animation of
our bottom tab bar.  In doing this, we fix a few weird behaviors that have
been observed since its introduction.

Instead, the new behavior animates the bottom tab bar based on actual
nested scroll events coming from the CoordinatorLayout.

Note: This contains a slight workaround for a known support library issue:
https://code.google.com/p/android/issues/detail?id=222911
This is done by adding `focusableInTouchMode1 to the FrameLayout that
contains our FeedView, to prevent it from automatically jump-scrolling on
certain devices.

Bug: T156012
Bug: T146890
Bug: T148689
Change-Id: I6707637ca228c732e901de4125094f2fd87f8578