Use G50 color for snackbar actions.
Our current action button text in Snackbars is a bit too low-contrasty.
This updates the color to G50.

It doesn't seem to be possible to style snackbars via xml, but
conveniently enough, we already dig into the underlying view hierarchy of
the Snackbar, so it's only a couple more lines of code to change the
action color programmatically.

Bug: T158898
Change-Id: I02fa1a9cab1fd5c4d17e2cd25956eaab31de5990