Don't crash when system WebView is being updated.
This patch makes PageActivity catch exception(s) related to inflation of
the WebView, meaning that we will now handle the condition when the app is
launched at the precise moment when the system WebView is being updated.

The behavior will be to show the user a brief message and immediately
close PageActivity. (but it's better than crashing.)

Change-Id: I6b2e2821f1df1117b90a5661901063a4b488895e