Make SnackBars appear above BottomNavigationView.
This is done by adding code to our custom Behavior that we use for the
BottomNavigationView. It now intercepts event when a Snackbar is added to
the CoordinatorLayout, and gives the Snackbar a bottom margin equal to
the height of the BottomNavigationView.  This means that we will now
explicitly show the bottom bar when a Snackbar is added, and prevent
hiding of the bottom bar for the duration of a Snackbar being shown.

(The Support library doesn't seem to have provisions to handle this
automatically.)

Bug: T159566
Change-Id: Ic56fc3a46bf1e9a3e309f85e86dbfc8bb889025d