Create custom view for page list items for use in RecyclerViews.
This introduces a custom view that represents a general-use page list item
that can be reused in any list of page titles, e.g. reading lists,
history, search results, etc.

This also adds an "action button" to each list item, which is hidden by
default, but can be enabled for any consumer that needs it.

Bug: T156073
Change-Id: Ia499b8cfd4902ea15af316c8250b7aa1d2c7e21c