Add info header to reading list fragment.
This is the header that shows the reading list title, description, and
article count at the top of the actual list of articles.

This is done by reusing the custom ReadingListItemView which is used for
showing the list of lists. We extend this view to accommodate a bit of
customization required for this purpose.

This header view is actually shown as another item in the RecyclerView,
but with a different "type", which distinguishes it from the page item
views.

Bug: T156073
Change-Id: I70be15f0be88a609a6426ce8b7fb222bd0bce26d