Use adjustResize, not adjustPan.
In MainActivity, invoking the soft keyboard doesn't actually resize the
screen contents to fit within the new viewport with the keyboard shown.
This means that cretain portions of the screen are inaccessible when the
keyboard is shown. (For example, go to History and click the Search icon
to show the keyboard, then try to scroll to the bottom. The last few items
in the list will be obscured by the keyboard.)

This fixes the issue by giving the activity a flag of "adjustResize"
instead of "adjustPan". We also set this flag programmatically for the
LanguagePreferenceDialog.

This does NOT address the issue in PageActivity, because the adjustResize
flag actually has no effect for full-screen activities with a WebView.
This is a known issue that is "working as intended":
https://code.google.com/p/android/issues/detail?id=5497

Bug: T130407
Change-Id: I0a26b28660d7ff44f9b4abdf0157f13348e3545b