Centralize/simplify add-to-reading-list snackbar.
This snackbar can actually be owned by the AddToList dialog itself, and
doesn't need to be callback'd to the originating activity.

This has the added benefit of automatically providing the snackbar for any
activity that adds items to a reading list, for example News (which
wasn't actually showing a snackbar at all until now).

Change-Id: If1b29f52362f4440448fc85fd5c031390e151d51