Enable deleting of reading list from reading list activity.
This fills in the TODO of deleting the whole reading list from the
individual reading list activity.

This is now done by invoking MainActivity and passing it an intent that
contains the title of the reading list to delete.  This title gets picked
up by ReadingListsFragment, and gets deleted via the normal lifecycle of
this fragment.

Change-Id: I00f6906b3cbb298a36e801e617f228487aed8896