Prevent possible crash when rendering feed card.
In CardLargeHeaderView, we animate the background color of the view to
match the dominant color of the image.
The patch of d67f70027a63 seems to have incorrectly modified the attribute
on which this animation is keyed.

This patch fixes the animator, and gives it an arguably more compatible
attribute on which to key the animation.

This patch also adds a check for whether the view is attached to a window
before starting the animation.

Bug: T160299
Change-Id: Ic1a40209189c905dea0f8c4d61cfe6412a674692