Prepare feed coordinator for improved event logging.
This improves the way that FeedCoordinator communicates back with
FeedFragment. Most importantly, we remove the update() callback, and use
only the insert() and remove() callbacks. This means that we will no
longer need to call notifyDataSetChanged(), which is great because it
messes with the view hierarchy of the RecyclerView, and could interfere
with our usage of onAttachedToWindow() for eventlogging purposes.

(i.e. when we call notifyDataSetChanged(), all the views get detached()
and re-attached(), which would cause a spurious event.)

Bug: T149125
Change-Id: Icc54fea1ed997151d34a2ad98efad0ebd3f4cee2