Bring back link preview overlay buttons.
This reenables the "Read article" button (and the "Get directinos" button
when in Nearby) to float translucently on top of the link preview dialog.

This is now accomplished by making the overlay into a custom view, and
injecting it into the view hierarchy at runtime when the dialog is shown.
The injection is done in onResume(), since it can't be done in
onCreateView() because the view hierarchy isn't constructed yet. We
search for a parent view called "id/content" (common to all dialogs)
which, conveniently, is a FrameLayout, and lets us float the overlay at
the bottom.

Bug: T159571
Change-Id: I781e259abb6fd08a4396b259da6111557f8dc250