Hide bottom navigation when ActionMode active.
This does a bit of rearranging of the view hierarchy in MainFragment which
allows us to finally get rid of the custom Behavior that we were using to
properly align SnackBars, as well as allow us to conditionally hide the
bottom tabs.

The tabs will now be hidden whenever an ActionMode is active (e.g. when
searching History or Lists).

Bug: T160636
Change-Id: I15418c8e2fc3c739cd976aead70ed7284f357d44