Use correct IME flag in PageActivity.
We were using the flag stateHidden, which would leave open the possibility
of the keyboard being shown even when the user isn't focused on an
editable field.

This changes the flag to stateAlwaysHidden, meaning that the keyboard will
always be hidden unless the user is explicitly focused on an editable
component.

Bug: T160694
Change-Id: I08baa08cb63c7369231669ec85a8e1e1c87b135d