Fix getting mobile authority for WikiSite when authority is already mobile
When instantiating WikiSite with a mobile domain, e.g. en.m.wikipedia.org,
and then requesting mobileAuthority() from the WikiSite, we were
incorrectly appending another "m." onto the domain that already has a "m."
in it.

This adds a check for whether the WikiSite's domain is already mobile
before appending the "m.", and adds some tests.

Bug: T161113
Change-Id: I0c907617df21a9d9ee8cc28ee971c8c36950b6c2