Introduce language blacklist for editing descriptions.
Instead of having a hard-coded whitelist of languages for which we enable
editing descriptions, we now have a blacklist which draws from a remote
config parameter, so that we'll be able to modify it without needing to
re-release the app.

In the current phase of our rollout, we're blacklisting the top ten wikis,
based on pageviews from within the app.

Change-Id: I84c0972bbf132150c3f4fa55910ee56625ddf316