Synchronize reading lists to user's account.
This synchronizes the logged-in user's reading lists to their
account using userjs options. This can effectively serve as a stopgap
measure until we have access to an authenticated API that is better than
userjs options, in the sense that they don't get passed with every
pageview when browsing in mobile web/desktop.

This patch aims to make it simple to switch out the backend storage API
for sending/receiving the remote set of lists, while keeping the same
structure for serializing the lists for storage.

The serialized structure itself is basically a minimization of what it
takes to construct a list. Crucially, the structure contains a `rev`
parameter that the client uses to determine whether its local state is out
of sync with the server.

This patch includes a hard-coded limit on the number of lists/pages that
can be synchronized, so that the user's userjs options don't grow out of
control.

TODO: Onboard the user that it's necessary to log in to their account for
their reading lists to be synced.

TODO: Hint to the user that only the first X pages/list will be synced.

Bug: T120107
Bug: T127702
Bug: T162763
Change-Id: I707d8855742edf2b27fca26020e59c6a58dd81f3