Don't prevent SavedPageSyncService from running when offline
Currently we stop the SavedPageSyncService from running when the device
is offline.  With the new caching setup, this is no longer always what we
want; it may well be the case that the content the user wants to save is
in the net cache and available for persisting in the save cache.  So, this
removes the check for online status.

This also allows the service to perform deletes while offline, which were
also prevented while offline before as a side effect of the check.

Hygiene: Removes the now-unused DeviceUtil.isOnline() method.

Bug: T162270
Change-Id: If98f58fe08324c195aec890d5f3cb976d08870a3