Explicitly declare which user options are synchronized.
This patch adds an explicit list that declares which userjs options are
synchronized to the user's account.

We've recently introduced a reading list synchronizer which uses a userjs
option for its own synchronization. But the thing is, our general
UserOption sync adapter reads *all* options that start with "userjs" and
upserts them into our database. This means that we'll have a duplicate
serialized reading list userjs option living in the database unnecessarily.

This patch simply limits which options are stored in the db (to the ones
we actually care about). This will prevent reading lists, as well as other
userjs cruft from unrelated MediaWiki gadgets and extensions, from being
stored in our local db.

Note: for good measure, I recommend clearing app preferences when
testing/merging this.

Change-Id: Id3fd417eed49b13fab0faf84eda1fe350eafc4d7