Fix: add alternate casing for device feature requirements
The app and Mapbox SDK's[0] AndroidManifests correctly mark the
following features optional:

- android.hardware.location
- android.hardware.location.gps
- android.hardware.wifi

The casing matches the documentation[1] but the Google Play Developer
Console is reporting that the following new features have been added:

- android.hardware.LOCATION
- android.hardware.location.GPS
- android.hardware.WIFI

It's not clear if these features are respecting the optional flag but
the number of devices supported is now far fewer. This patch adds
additional redundant uses-feature nodes that match the casing used in
the Console. This patch should be reverted if it does not fix the
complaint. However, other users appear to be experiencing similar
issues.[2]

[0] https://github.com/mapbox/mapbox-gl-native/blob/62ea1f21858c69f6921c775ba7a3de201f0514d8/platform/android/MapboxGLAndroidSDK/src/main/AndroidManifest.xml#L5
[1] https://developer.android.com/guide/topics/manifest/uses-feature-element.html
[2] http://stackoverflow.com/questions/14299956/i-set-uses-feature-androidrequired-false-but-google-play-keeps-insisting-on-t

Change-Id: Ib9f21dab6ac2e84a39e502f47732739ea2563af2