Upload user options only when they're modified.
Looking closely at logcat, I noticed that we were always uploading user
options unconditionally every time they were synced. This would send
an unnecessary POST request for every option, whenever a sync is
requested.

This patch updates the logic so that each option is POSTed only if it's
not yet marked as synchronized in the db.

Change-Id: Ia317dafb1c04df741f0087b9d465c8b1601d8c34