Fix: prevent race condition when synchronizing user options.
Some of our recent changes (probably the refactor of CSRF token fetching)
seems to have introduced a race condition in the synchronizing logic of
user options.  This is likely because the method to post() a user option
became asynchronous, whereas it was previously synchronous.

This creates a condition where we request to post() an updated user
option, then immediately call get() to retrieve the latest options from
the server, when in fact the previous post() call hasn't yet finished.

This patch updates the post() and delete() methods to be 'artificially'
synchronous by using wait() / notify() logic.

TODO: update synchronizing of user options to use a 'revision' code,
similar to how reading lists are synchronized.

Bug: T163214
Change-Id: Ia8c7ac33d9cb52fad0d1676445f419a573292de1