Stop thrashing on page save failure and save on subsequent runs
Stops relying on a ContentObserver to trigger the saved page sync service.
This isn't viable because of the way the ContentObserver works and the way
the DAO works.

Instead, we'll manually kick off saved page syncing after a reading list
sync event (such as happens when the app is launched or after a page is
added to a reading list).

No more thrashing!

Also added logging for successful page saves for dev convenience.

Bug: T162894
Change-Id: I1d0bfd2e2071942f3c7ad977ae271cf78db82235