Crash fix: Remove reliance on a Context to detect a 404
This should fix the top crash from the 4/19 beta:

https://rink.hockeyapp.net/manage/apps/226650/app_versions/93/crash_reasons/165970817

Updates ThrowableUtil.is404 to drop its alternate branch which draws upon
ThrowableUtil.getAppError.  404s are only expected in the context of
missing pages or summaries for them, both of which we get with Retrofit,
not java-mwapi.  Retrofit throws HttpStatusExceptions, which we don't
need a Context to check.  If we get 404s from the MediaWiki API (as
opposed to the expected 200s with error messages), we've got bigger
problems, and in any case, java-mwapi will be gone Real Soon Now (TM).

Perhaps more controversially, deprecates getAppError and its supporting
methods as well as the AppError class itself.  Our philsophy around user
error messages has shifted since this code was written in favor of
standard, predefined error views. Using these legacy methods requires
passing in a Context, which can be hazardous.  Further, they're already
nearly unused.

Change-Id: I53c944b2b96ac2ca9e36191b582fb13d5f5c2455