Allow backup of app database.
This allows the system to back up our app's database to the user's Google
account.[1] This means that if the user performs a backup of their device,
and then uninstalls/reinstalls our app (or resets their device), the
database should be restored automatically. The user will then regain
access to their previously-saved reading lists and browsing history, which
at least partially addresses our users' needs for reading lists to be
persisted across app installs.

Note that this instructs the system to ONLY back up the database, and not
any of the app's other files, such as the contents of saved pages, which
can be quite large, and exceed Google's 25MB limit for app backups.

This also does NOT back up the app's SharedPreferences, which can contain
the user's login credentials, and could present a security issue if backed up.

For testing the backup functionality, see [2].

[1] https://developer.android.com/guide/topics/data/autobackup.html
[2] https://developer.android.com/guide/topics/data/testingbackup.html

Bug: T130331
Change-Id: Ie659913abd51e6312571e281842707a5fcc0eec6