Fix: Centralize retrieval of DB name from WikiSite.
The "dbname" is yet another alias for a wiki site that is used for certain
API calls.  Usually this is represented by the language code plus the word
"wiki", such as "enwiki", "dewiki", etc.

However, this is not true for *all* language codes. For instance, the
dbname for Norwegian Bokmal Wikipedia is "nowiki", even though the
language code is "nb".  There may be several more of these cases, but they
can be added later.

This patch creates a centralized method where the dbname can be retrieved
directly from a given WikiSite, and adds some tests.

Bug: T163920
Change-Id: If4d33bb8038c00fbae914e5ff00148ccb9cceeeb