Fix: Update MCS content format versions accepted
It's important to keep the MCS content versions requested by the current
development version of the app in sync with the latest MCS output version.

In this case, our failing to update our accept header with the latest
mobile-sections version meant that the app was not benefiting from a fix
deployed nearly a month ago, on April 10.

Problems around this probably haven't shown up until now because lately
Bernd is being more proactive with version bumps.

I watch the MCS repo reasonably closely so I'll make sure to keep this up-
to-date going forward.

Bug: T162116
Change-Id: I8806001caeadd2b6bd28207fbcf5c329e0dd3dd7