Improve preserving of backstack when saving instance state.
When saving PageActivity's instance state, and indeed when pausing this
activity for any reason, we persist the current backstack of PageTitles,
so that when the activity reloads, we load the last item from the
backstack.

However, we were not persisting the description or thumbUrl of the
PageTitles in the backstack, since these are always loaded dynamically
when we load the page content into the WebView.  The problem is when
another component starts depending on a title from the backstack *before*
the page content is loaded.  An example is the description editing
activity. This activity could be launched before PageActivity has a chance
to load the page content, and can therefore receive a PageTitle with an
empty description, which can lead to the behavior seen in the phab task.

This patch simply updates the additional metadata of the description and
thumbUrl onto the current backstack entry when pausing the activity.

Bug: T157824
Change-Id: I9700c60185d9ca9147f15bb2fedcf853a3b48d27