Fix: use proper margins in description editing view for API 16.
API 16 doesn't actually support marginStart/marginEnd, although the IDE
doesn't show a warning about this when they're specified together.

This changes our usage of margins in the description edit view to the more
compatible marginLeft+marginRight. This seems to be the only place where
this is the case. In other places where we use a single Start or End
margin, it's accompanied by an equivalent Left or Right margin.

This also removes an unnecessary margin from the About view (looks like a
holdover from when the W icon was inline with the Wikipedia wordmark).

Change-Id: If7e859cfcaac54980f9e5a2051e661d4ae546617