Stop updating the feed UI based on ReadingListPageObserver reports
Reverts the logic added to observe changes on the reading list page disk
table and report them back to the feed to update UI.  It resulted in
little gain and caused lots of problems, including a flickering feed UI
and jumping context menus.  Featured article card footers will still be
updated immediately when changed from the feed itself.

This makes ReadingListPageObserver obsolete, so it's removed.

Bug: T164374
Change-Id: I383a8f7d40016659bff43895ea06c941957d004c