Render math formulas correctly on API <19.
On API 18 and below, our math formulas were being rendered very weirdly,
and would usually appear as very tiny images.

After much investigating, it turns out this is because of the old
WebView's incomplete support for SVG graphics.  Specifically, the SVG
files that are used for our math formulas make use of "ex" units when
specifying their nominal width and height. The WebView does not seem to
handle these units correctly.

This patch makes use of our logic that intercepts requests made by the
WebView, and checks if the content-type of the response is SVG.  And if
it's SVG, we rewrite any usages of "ex" units with "em".  This fixes the
display of formulas on older APIs, and doesn't seem to have any negative
effects on other non-mathematical SVGs.

Bug: T107775
Change-Id: Icee31fcc16b4addee9c4a328c01cc88f7b0caccf