Fix possible memory leaks when destroying WebViews.
Observed (and fixed) in at least APIs 18 and 22:
There seem to be specific circumstances when the WebView holds on to
certain View references, which in turn hold on to an Activity reference
indefinitely, causing a leak.

For example, if we navigate to the [[Obama]] article and then exit the
activity, the WebView holds on to an instance of `HTML5VideoView`, which
is presumably a container for a video object that's shown in the article.

There's a lot of conflicting information and workarounds for this on the
web, but a common thread seems to be that the WebView doesn't always
destroy its own view hierarchy correctly, and needs to be triggered
explicitly.

This patch adds a method to our ObservableWebView that does its best to
destroy itself, which includes:
- Clearing all listeners
- Removing itself from the Activity's view hierarchy
- Calling destroy() on itself, which does the actual work of
  uninitializing the WebView's internals, and relies on the WebView no
  longer being attached to a window.

Change-Id: I6dc550eec706c7a2041cdad61d55c00e4558c52d