Trigger saving/unsaving of page upon toggle.
The user interaction of toggling the saved state of a page wasn't actually
kicking off the sync service to update the page's state in the database
and on disk.

Technically the service will still be kicked off the next time the
activity is launched, but this patch makes it immediate.

Bug: T164585
Change-Id: I0bbb61e3d3ff04e8b9501c01cbeae2dfbc7d2124