Do not assume that all lead image thumb URLs are without scheme
Until now, the app expects to get a thumb image URL returned from the API,
which does not contain a scheme. This isn't true, at least not for all
third-parties, where probably a https URL is returned through the API. The
app now handles these URLs by parsing the returned URL to an URI object
and sets the scheme from the WikiSite only, if there's no scheme in the
URL already.

Until now, the current implemented handling would result in stupid things
like:

  httpshttps://uploads.wikipedia.org...

Bug: T165495
Change-Id: I96d624a3bdfe0c76890a1993dc31eb3af600a1bc