Added MediaWiki flavour
Since, we are not using any suffix here, we had to modify the appId
function in build.gradle in order to avoid a conflict with the
provider authority.

Modified some drawables and strings to use the MediaWiki ones. This
is still work in progress for other subtask.

Refactored the activity_about layout in order to avoid code duplication.
Corrected strings for the MediaWiki specific flavour.

Replaced Wikipedia in strings.

Bug: T165783
Change-Id: Ie1df1f17668a40ad1c1101fce9f21beae213575c