Update: Sort search results in the SearchResults constructor
Search results sorting was omitted, by me (facepalm), from the recent
patch moving full text search queries to Retrofit.  It was partially
restored by Jcasariego in a downstream consumer.  This centralizes sorting
in the new SearchResults constructor, so that it doesn't have to be
duplicated either in both search clients or in all n consumers.

Also takes out the recent addition of an 'index' property to PageTitle
and restores it to being solely a concern of MwQueryPage.

Change-Id: I154e6730573782a98ba0bc2bcf6cf8291fc61eab