[Upgrade fixes] Prevent crash when upgrading from 2 versions ago.
If the app is upgraded directly to the current version from a previous
version prior to the introduction of the saved-page cache delegate, it
crashes because the SavedPageSyncService wants to immediately call Stat on
the cache directory, when the directory doesn't exist yet (before any new
network calls are made).

This patch makes sure the cache directory gets created upon instantiating
the cache delegate, so that we can call Stat on it even before any network
requests are called.

Change-Id: I1e79fa7350fcaebc58120786417648cf703ee8ba