Fix request date/"age" handling when coming from OfflineCard
The biggest change introduced here is that the pendingClients list is a
bit more stateful.  The current FeedClient remains at pos 0 in the
pendingClients list until it is removed when requestNextCard() is called,
so that a request can be retried when coming from offline.

This removes the need for brittle and probably fundamentally incorrect
things like resetting the age back a day when offline state is detected.

Bug: T166789
Change-Id: I83147fe5a49a77802ab8f887233ce15cd6b98d90