Revert "Only request Feed content if RESTBase is enabled"
I79e20a8a exposed a race condition in which, on a fresh install, the
feed content is likely to be requested and layout begun before the remote
config is fetched.  I79e20a8a added a check that RESTBase loading was
enabled in order for RB-based feed content to be loaded, so it is likely
that, on first launch on a fresh install, only the main page and random
cards (the only two fully local cards) will appear in the feed, since the
remote config variable permitting RESTBase loading will not yet have been
fetched.  When the device is online, the remote config will likely be
fetched by the time the user attempts a refresh, useRestbase will then be
true, and the expected feed content will appear.

When we're offline, however, the remote config fetch call will fail, and
will not be attempted for another 24 hours.  In the meantime, the user
will get no feed content because useRestbase will be false.  This
shouldn't happen.

Really, the race condition should be fixed.  I'll file that as a follow-
up task.  But, the useRestbase restriction on loading feed content seems
unnecessary in any case, so this reverts it.

This reverts commit 204810b1a3519f8c89f77e90886d8efb6669b5eb.

Bug: T166789
Bug: T166922
Change-Id: I19e1032545485b31820ca64f46e4d304f755721e