Consolidate possible MW API query result components in MwQueryResult
Removes the type variable from MwQueryResponse and creates a new MwQuery-
Result class where all of the possible components that we use reside.

This removes the need to specify unwieldy type parameters in every data
client and consumer.

More importantly, it provides for query responses containing multiple
result types, such as are used in TitleSearchTask.

Change-Id: I8845f8e6342b3af0fc59ed5c83c0619da48641e0