Fix possible crash when saving page with malformed image URL(s).
The outer exception handler for saving a page was only expecting
IOException, when in fact the network call may throw additional exceptions
like IllegalArgumentException if the URL is malformed.

https://rink.hockeyapp.net/manage/apps/226650/app_versions/102/crash_reasons/172373534

This also tweaks the logic of the page saving loop, whereby an exception
caused by a single page *continues* the loop instead of breaking out of
it, preventing conditions such as malformed URLs from blocking the
download of the remaining queue of pages.

Change-Id: I10c46411ea0e26b67d2f78aed0b96eb95961c748