Fix: Update search results cache with full text search results
The code for processing full text search results has long contained a
code snippet that was intended to update the cached search results from
the preceding title prefix search with the new full text search results.
What was actually happening is that a List<SearchResult> object was created
and the new results were appended to it, and then nothing was done with
it.

This adds a line to re-cache the updated results.

Change-Id: I86b71b707204ad60efc7f931ae74c64b69b9396f