Update okhttp to 3.8.0
N.B. Involves a change to a modified bit of CacheDelegateInterceptor!
Please review carefully and test well!  Here's the relevant okhttp commit:

Invalidate the cache even if the response has no body.
https://github.com/square/okhttp/commit/c9496d350c0c3bb4b9d9571c3ab97181726561e2

Also of note:

Version 3.8.0

New: The response message is now non-null. This is the "Not Found" in the
status line "HTTP 404 Not Found". If you are building responses programma-
tically (with new Response.Builder()) you must now always supply a
message. An empty string "" is permitted. This value was never null on
responses returned by OkHttp itself, and it was an old mistake to permit
application code to omit a message.

Version 3.7.0

New: Connection coalescing. OkHttp may reuse HTTP/2 connections across
calls that share an IP address and HTTPS certificate, even if their domain
names are different.

New: MockWebServer's RecordedRequest exposes the requested HttpUrl with
getRequestUrl().

https://github.com/square/okhttp/blob/master/CHANGELOG.md

Change-Id: Ie2cc3bf806b7a6c45bc31a068fa4c64cecad51de