Use AccountManager for storing user credentials.
This transitions our code to fully utilize AccountManager to store our
user credentials. We were actually already halfway-using AccountManager;
this just takes it all the way there.
Some additional highlights:

- All account management functions are in the static AccountUtil class,
  which implicitly expects there to be at most one account for our app in
  the system. It's possible that we might want to support multiple
  simultaneous logged-in accounts in the future, but certainly not yet.
- Remove the User and UserInfoStorage classes.
- Provide a one-time migration function that makes sure to create an
  AccountManager account based on the old Prefs, and then delete the prefs
  permanently.

Note on the new behavior of this logic: The user's account will now
survive the clearing of the app data in Settings!

Bug: T165357
Change-Id: I1956657a27c7cab4c33938b42954861f5087ae5b